cpufreq: Add Smartmax governor
Smartmax is a mix between ondemand and smartassv2. It behaves mostly like smartass with the concept of an "ideal" frequency. By default this is configured for battery saving, so this is NOT a gaming or benchmark governor! Additionally, to make it "snappy", smartmax has "touch poke". So input events from the touchscreen will boost the cpu for a specific time to a specific frequency. Developed by XDA user Maxwen.

Change-Id: If1817467f2218cc2d48b900af39b70a0ffd8ccdb
Signed-off-by: maxwen <max.weninger@gmail.com>